Correct linux build dependencies for Depth Quality tool Remove redundant files Continue realsense-viewer refactoring
(cherry picked from commit be744450cdfc640759fc632cfbb1808225daaf30)

+++ common/rendering.h
@@ -292,8 +292,8 @@ namespace rs2
                 H *= scale;
             }
 
-            return{ floor(x + floor(w - W) / 2), 
-                    floor(y + floor(h - H) / 2), 
+            return{ float(floor(x + floor(w - W) / 2)),
+                    float(floor(y + floor(h - H) / 2)),
                     W, H };
         }
